Support prefix strings that start with a vowel
Because the string devoweler made a substring searching from the left which
broke on the first occurrence of a vowel (non-inclusive), it would return an
empty string if the prefix began with a vowel. For example, 'Amy' as a prefix
string would return ''.

This commit preserves the first letter of the prefix string, whether or not
it's a vowel.